requirements_pypi.txt
###### This requirements file is used for local installation where access to GitHub is restricted ######
###### Common requirements ######
lxml >= 4.3.0
six (>=1.12)
requests (>=2.27.1)
pyOpenSSL (>=22.0.0)
cryptography (>=37.0.2)
setuptools (>=62.3.2)
###### SDK requirements ######
pyVmomi==9.0.0.0
vmware-vapi-common-client==2.61.2
vmware-vapi-runtime==2.61.2
vmware-vcenter == 9.0.0.0
vmware-vsan-data-protection==9.0.0.0